1.1 layout (#1)
* basic logic

need the 4 images from the center

* home now ready

only need a nav bar
