--- IndexView.php
@@ -1,36 +1,0 @@
-<?php 
-require_once("Autoloader.php");
-class IndexView
-{
-	private $IndexController;
-	private $IndexModel;
-
-	public function __construct($indexController, $indexModel)
-	{
-		$this->IndexController = $indexController;
-		$this->IndexModel = $indexModel;
-	}
-
-	//output to html
-	public function output(){
-		$page = "";
-		$page .= $this->Header();
-		$page .= $this->Body();
-		$page .= $this->Footer();
-		return $page;
-	}
-
-	private function Header(){
-		return $this->IndexController->GetConfig()->GetHeader("Index");
-	}
-
-	private function Body(){
-		return "<div id='main'>
-		 L Lorem ipsum dolor sit amet, consectetuer adipiscing elit. Aenean commodo ligula eget dolor. Aenean massa. Cum sociis natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus. Donec quam felis, ultricies nec, pellentesque eu, pretium quis, sem. Nulla consequat massa quis enim. Donec pede justo, fringilla vel, aliquet nec, vulputate eget, arcu. In enim justo, rhoncus ut, imperdiet a, venenatis vitae, justo. Nullam dictum felis eu pede mollis pretium. Integer tincidunt. Cras dapibus. Vivamus elementum semper nisi. Aenean vulputate eleifend tellus. Aenean leo ligula, porttitor eu, consequat vitae, eleifend ac, enim. Aliquam lorem ante, dapibus in, viverra quis, feugiat a, tellus. Phasellus viverra nulla ut metus varius laoreet. Quisque rutrum. Aenean imperdiet. Etiam ultricies nisi vel augue. Curabitur ullamcorper ultricies nisi. Nam eget dui.</div>";
-	}
-
-	private function Footer(){
-		return "</body></html> ";
-	}
-}
-?>
