+++ Style.css
@@ -1,0 +1,60 @@
+.RedBar{
+	width: 100%;
+	height: 20px;
+	background-color: #DD042B;
+	margin-bottom: 5px;
+}
+
+.Logo{
+	text-align: right;
+	width: 550px;
+	margin-left: 50px;
+	margin-top: 50px;
+}
+
+.HomepageImages{
+	height: 700px;
+}
+
+.HomepageImage{
+	margin-right: -240px;
+	width: 500px;
+	clip-path: polygon(42% 0%, 92% 0%, 55% 100%, 5% 100%);
+}
+
+.HomepageImage:hover{
+	margin-right: -240px;
+	width: 500px;
+	opacity: 50%;
+}
+
+.lol{
+	margin-top: 90px;
+}
+
+.rotate {
+    -webkit-transform: rotate(45deg);
+    -moz-transform: rotate(45deg);
+    -o-transform: rotate(45deg);
+    -ms-transform: rotate(45deg);
+    transform: rotate(45deg);
+}
+
+.MediaIcons{
+	float: right;
+	margin-right: 5px;
+	width: 30px;
+	height: 30px;
+}
+
+.Footer{
+	margin-top: 10px;
+	border: solid 1px;
+	height: 45px;
+}
+
+#DesignedBy{
+	margin-top: 6px;
+	margin-left: 30%;
+	font-size: 20px;
+}
